requirements.txt
pandas
colorlog
pycodestyle
packageurl-python
pylint
black
